moved most everything
