+++ web/api/webrpc/pipeline.go
@@ -4,16 +4,13 @@ import (
 	"context"
 	"time"
 
-	"github.com/filecoin-project/curio/lib/curiochain"
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-bitfield"
 	"github.com/snadrus/must"
 	"golang.org/x/xerrors"
 
-	"github.com/filecoin-project/lotus/blockstore"
 	"github.com/filecoin-project/lotus/chain/actors/adt"
 	"github.com/filecoin-project/lotus/chain/actors/builtin/miner"
-	"github.com/filecoin-project/lotus/chain/store"
 	"github.com/filecoin-project/lotus/chain/types"
 )
 
@@ -105,7 +102,6 @@ func (a *WebRPC) PipelinePorepSectors(ctx context.Context) ([]sectorListEntry, e
 
 	minerBitfieldCache := map[address.Address]minerBitfields{}
 
-	stor := store.ActorStore(context.Background(), blockstore.NewReadCachedBlockstore(blockstore.NewAPIBlockstore(a.deps.Chain), curiochain.ChainBlockCache))
 	sectorList := make([]sectorListEntry, 0, len(tasks))
 	for _, task := range tasks {
 		task := task
@@ -119,7 +115,7 @@ func (a *WebRPC) PipelinePorepSectors(ctx context.Context) ([]sectorListEntry, e
 
 		mbf, ok := minerBitfieldCache[addr]
 		if !ok {
-			mbf, err = a.getMinerBitfields(ctx, addr, stor)
+			mbf, err = a.getMinerBitfields(ctx, addr, a.stor)
 			if err != nil {
 				return nil, xerrors.Errorf("failed to load miner bitfields: %w", err)
 			}
@@ -188,4 +184,63 @@ func (a *WebRPC) getMinerBitfields(ctx context.Context, addr address.Address, st
 		unproven:  unproved,
 		faulty:    faulty,
 	}, nil
+}
+
+type PorepPipelineSummary struct {
+	Actor string
+
+	CountSDR          int
+	CountTrees        int
+	CountPrecommitMsg int
+	CountWaitSeed     int
+	CountPoRep        int
+	CountCommitMsg    int
+	CountDone         int
+	CountFailed       int
+}
+
+func (a *WebRPC) PorepPipelineSummary(ctx context.Context) ([]PorepPipelineSummary, error) {
+
+	head, err := a.deps.Chain.ChainHead(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	rows, err := a.deps.DB.Query(ctx, `
+	SELECT 
+		sp_id,
+		COUNT(*) FILTER (WHERE after_sdr = false) as CountSDR,
+		COUNT(*) FILTER (WHERE (after_tree_d = false OR after_tree_c = false OR after_tree_r = false) AND after_sdr = true) as CountTrees,
+		COUNT(*) FILTER (WHERE after_tree_r = true and after_precommit_msg = false) as CountPrecommitMsg,
+		COUNT(*) FILTER (WHERE after_precommit_msg_success = true AND seed_epoch > $1) as CountWaitSeed,
+		COUNT(*) FILTER (WHERE after_porep = false AND after_precommit_msg_success = true) as CountPoRep,
+		COUNT(*) FILTER (WHERE after_commit_msg_success = false AND after_porep = true) as CountCommitMsg,
+		COUNT(*) FILTER (WHERE after_commit_msg_success = true) as CountDone,
+		COUNT(*) FILTER (WHERE failed = true) as CountFailed
+	FROM 
+		sectors_sdr_pipeline
+	GROUP BY sp_id`, head.Height())
+	if err != nil {
+		return nil, xerrors.Errorf("query: %w", err)
+	}
+	defer rows.Close()
+
+	var summaries []PorepPipelineSummary
+	for rows.Next() {
+		var summary PorepPipelineSummary
+		var actor int64
+		if err := rows.Scan(&actor, &summary.CountSDR, &summary.CountTrees, &summary.CountPrecommitMsg, &summary.CountWaitSeed, &summary.CountPoRep, &summary.CountCommitMsg, &summary.CountDone, &summary.CountFailed); err != nil {
+			return nil, xerrors.Errorf("scan: %w", err)
+		}
+
+		sactor, err := address.NewIDAddress(uint64(actor))
+		if err != nil {
+			return nil, xerrors.Errorf("failed to create actor address: %w", err)
+		}
+
+		summary.Actor = sactor.String()
+
+		summaries = append(summaries, summary)
+	}
+	return summaries, nil
 }
