+++ web/api/webrpc/routes.go
@@ -8,8 +8,13 @@ import (
 
 	"github.com/filecoin-project/go-jsonrpc"
 
+	"github.com/filecoin-project/lotus/blockstore"
+	"github.com/filecoin-project/lotus/chain/actors/adt"
+	"github.com/filecoin-project/lotus/chain/store"
+
 	"github.com/filecoin-project/curio/build"
 	"github.com/filecoin-project/curio/deps"
+	"github.com/filecoin-project/curio/lib/curiochain"
 )
 
 var log = logging.Logger("webrpc")
@@ -17,6 +22,7 @@ var log = logging.Logger("webrpc")
 type WebRPC struct {
 	deps      *deps.Deps
 	taskSPIDs map[string]SpidGetter
+	stor      adt.Store
 }
 
 func (a *WebRPC) Version(context.Context) (string, error) {
@@ -30,6 +36,7 @@ func (a *WebRPC) BlockDelaySecs(context.Context) (uint64, error) {
 func Routes(r *mux.Router, deps *deps.Deps) {
 	handler := &WebRPC{
 		deps:      deps,
+		stor:      store.ActorStore(context.Background(), blockstore.NewReadCachedBlockstore(blockstore.NewAPIBlockstore(deps.Chain), curiochain.ChainBlockCache)),
 		taskSPIDs: makeTaskSPIDs(),
 	}
