+++ web/static/index.html
@@ -14,6 +14,7 @@
     <script type="module" src="cluster-tasks.mjs"></script>
     <script type="module" src="cluster-machines.mjs"></script>
     <script type="module" src="cluster-task-history.mjs"></script>
+    <script type="module" src="pipeline-porep.mjs"></script>
     <style>
         .deadline-box {
             display: grid;
@@ -79,32 +80,7 @@
 
             <cluster-machines></cluster-machines>
 
-            <div class="row">
-                <div class="col-md-auto" style="max-width: 1000px">
-                    <div class="info-block">
-                        <h2>PoRep Pipeline</h2>
-                        <table class="table table-dark">
-                            <thead>
-                            <tr>
-                                <th>Address</th>
-                                <th>SDR</th>
-                                <th>Trees</th>
-                                <th>Precommit Msg</th>
-                                <th>Wait Seed</th>
-                                <th>PoRep</th>
-                                <th>Commit Msg</th>
-                                <th>Done</th>
-                                <th>Failed</th>
-                            </tr>
-                            </thead>
-                            <tbody hx-get="/hapi/simpleinfo/pipeline-porep" hx-trigger="load,every 5s">
-                            </tbody>
-                        </table>
-                    </div>
-                </div>
-                <div class="col-md-auto">
-                </div>
-            </div>
+            <pipeline-porep></pipeline-porep>
 
             <div class="row">
                 <div class="col-md-auto" style="max-width: 1000px">
