+++ web/static/pages/pipeline_porep/pipeline_porep_sectors.mjs
@@ -1,9 +1,12 @@
-import { html, css, LitElement } from 'https://cdn.skypack.dev/lit';
+import { LitElement, html, css } from 'https://cdn.jsdelivr.net/gh/lit/dist@3/all/lit-all.min.js';
 
 class PipelinePorepSectors extends LitElement {
     constructor() {
         super();
         this.data = [];
+    }
+
+    connectedCallback() {
         this.loadData();
     }
 
@@ -33,7 +36,46 @@ class PipelinePorepSectors extends LitElement {
         .porep-pipeline-table tr:first-child {
             border-top: none;
         }
+    `;
 
+    render() {
+        return html`
+            <link rel="stylesheet" href="/ux/main.css">
+            <style>${PipelinePorepSectors.styles}</style>
+            <table class="table table-dark porep-state porep-pipeline-table">
+                <tbody>
+                    ${this.renderSectors()}
+                </tbody>
+            </table>
+        `;
+    }
+
+    renderSectors() {
+        return this.sectors.map((sector) => html`
+            <tr>
+                <td>${sector.Address}</td>
+                <td rowspan="2">${sector.CreateTime}</td>
+                <td rowspan="2">${this.renderSector(sector)}</td>
+                <td rowspan="2">
+                    <a href="/pages/sector/?sp=${sector.Address}&sector=${sector.SectorNumber}">DETAILS</a>
+                </td>
+            </tr>
+            <tr>
+                <td>${sector.SectorNumber}</td>
+            </tr>
+        `);
+    }
+
+    static properties = {
+        sectors: { type: Array },
+    };
+}
+
+customElements.define('pipeline-porep-sectors', PipelinePorepSectors);
+
+// split renderSector to <sector-porep-state>
+class SectorPorepState extends LitElement {
+    static styles = css`
         .porep-state {
             border-collapse: collapse;
         }
@@ -67,71 +109,64 @@ class PipelinePorepSectors extends LitElement {
 
         .pipeline-failed {
             background-color: #603030;
-        }
-    `;
+        }`
+    properties = {
+        sector: Object,
+    }
 
     render() {
+        const sector = this.sector;
         return html`
-            <link rel="stylesheet" href="/ux/main.css">
-            <style>${PipelinePorepSectors.styles}</style>
-            <table class="table table-dark porep-state porep-pipeline-table">
+            <table class="porep-state">
                 <tbody>
-                    ${this.renderSectors()}
+                    <tr>
+                        ${this.renderSectorState('SDR', 1, sector.TaskSDR, sector.AfterSDR)}
+                        ${this.renderSectorState('TreeC', 1, sector.TaskTreeC, sector.AfterTreeC)}
+                        ${this.renderSectorState('PComm Msg', 2, sector.TaskPrecommitMsg, sector.AfterPrecommitMsg)}
+                        ${this.renderSectorStateNoTask('PComm Wait', 2, sector.AfterPrecommitMsg, sector.AfterPrecommitMsgSuccess)}
+                        ${this.renderSectorStateNoTask('Wait Seed', 2, sector.AfterPrecommitMsgSuccess, sector.AfterWaitSeed)}
+                        ${this.renderSectorState('PoRep', 2, sector.TaskPoRep, sector.AfterPoRep)}
+                        ${this.renderSectorState('Clear Cache', 1, sector.TaskFinalize, sector.AfterFinalize)}
+                        ${this.renderSectorState('Move Storage', 1, sector.TaskMoveStorage, sector.AfterMoveStorage)}
+                        <td class="${sector.ChainSector ? 'pipeline-success' : (sector.ChainAlloc ? 'pipeline-active' : 'pipeline-failed')}">
+                            <div>On Chain</div>
+                            <div>${sector.ChainSector ? 'yes' : (sector.ChainAlloc ? 'allocated' : 'no')}</div>
+                        </td>
+                        <td rowspan="2" class="${sector.Failed ? 'pipeline-failed' : (sector.ChainActive ? 'pipeline-success' : 'pipeline-active')}">
+                            <div>State</div>
+                            <div>${sector.Failed ? 'Failed' : (sector.ChainActive ? 'Sealed' : 'Sealing')}</div>
+                        </td>
+                    </tr>
+                    <tr>
+                        ${this.renderSectorState('TreeD', 1, sector.TaskTreeD, sector.AfterTreeD)}
+                        ${this.renderSectorState('TreeR', 1, sector.TaskTreeR, sector.AfterTreeR)}
+                        <!-- PC-S, PC-W, WS, PoRep -->
+                        ${this.renderSectorState('Commit Msg', 1, sector.TaskCommitMsg, sector.AfterCommitMsg)}
+                        ${this.renderSectorStateNoTask('Commit Wait', 1, sector.AfterCommitMsg, sector.AfterCommitMsgSuccess)}
+                        <td class="${sector.ChainActive ? 'pipeline-success' : 'pipeline-failed'}">
+                            <div>Active</div>
+                            <div>${sector.ChainActive ? 'yes' : (sector.ChainUnproven ? 'unproven' : (sector.ChainFaulty ? 'faulty' : 'no'))}</div>
+                        </td>
+                    </tr>
                 </tbody>
             </table>
         `;
     }
-
-    renderSectors() {
-        return this.sectors.map((sector) => html`
-            <tr>
-                <td>${sector.Address}</td>
-                <td rowspan="2">${sector.CreateTime}</td>
-                <td rowspan="2">${this.renderSectorState(sector)}</td>
-                <td rowspan="2">
-                    <a href="/hapi/sector/${sector.Address}/${sector.SectorNumber}">DETAILS</a>
-                </td>
-            </tr>
-            <tr>
-                <td>${sector.SectorNumber}</td>
-            </tr>
-        `);
-    }
-
-    renderSectorState(sector) {
+    renderSectorStateNoTask(name, rowspan, active, after) {
         return html`
-            <td class="${this.getClasses(sector.TaskSDR, sector.AfterSDR)}">
-                <div>SDR</div>
-                <div>${this.renderTask(sector.TaskSDR, sector.AfterSDR)}</div>
+            <td rowspan="${rowspan}" class="${active?'pipeline-active':''} ${after?'pipeline-success':''}">
+                <div>${name}</div>
+                <div>${after?'done':'--'}</div>
             </td>
-            <!-- Rest of the table cells... -->
         `;
     }
-
-    getClasses(task, after) {
-        let classes = '';
-        if (task !== null) {
-            classes += 'pipeline-active ';
-        }
-        if (after) {
-            classes += 'pipeline-success ';
-        }
-        return classes.trim();
+    renderSectorState(name, rowspan, task, after) {
+        return html` 
+            <td rowspan="${rowspan}" class="${task?'pipeline-active':''} ${after?'pipeline-success':''}">
+                <div>${name}</div>
+                <div>${after?'done':task?'T:'+task:'--'}</div>
+            </td>
+        `;
     }
-
-    renderTask(task, after) {
-        if (after) {
-            return 'done';
-        } else if (task !== null) {
-            return `T:${task}`;
-        } else {
-            return '--';
-        }
-    }
-
-    static properties = {
-        sectors: { type: Array },
-    };
 }
-
-customElements.define('pipeline-porep-sectors', PipelinePorepSectors);
+customElements.define('sector-porep-state', SectorPorepState);
