+++ web/static/pages/sector/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <title>Node Info</title>
+    <script type="module" src="/ux/curio-ux.mjs"></script>
+    <script type="module" src="/sector-info.mjs"></script>
+  </head>
+  <body style="visibility:hidden; background:rgb(11, 22, 34)" data-bs-theme="dark">
+    <curio-ux>
+      <section class="section">
+        <div class="row">
+          <h1>Sector Info</h1>
+        </div>
+      </section>
+      <section class="section">
+        <div class="row">
+            <div class="col-md-auto" style="max-width: 95%">
+            <sector-info></sector-info>
+            </div>
+          </div>
+        </div>
+      </section>
+    </curio-ux>
+</html>
